requirements.txt
2captcha-python==1.2.2
async-timeout==4.0.3
attrs==23.2.0
Automat==22.10.0
beautifulsoup4==4.12.3
boto3==1.34.42
botocore==1.34.42
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.2
cssselect==1.2.0
dnspython==2.5.0
exceptiongroup==1.2.0
filelock==3.13.1
h11==0.14.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
jmespath==1.0.1
lxml==5.1.0
outcome==1.3.0.post0
packaging==23.2
parsel==1.8.1
pillow==10.2.0
Protego==0.3.0
pyaes==1.6.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
pymongo==4.6.1
pyOpenSSL==24.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-socks==2.4.4
queuelib==1.6.2
requests==2.31.0
requests-file==2.0.0
rsa==4.9
s3transfer==0.10.0
Scrapy==2.11.0
scrapy-rotating-proxies==0.6.2
scrapy-selenium==0.0.7
selenium==4.9.1
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
Telethon==1.34.0
tldextract==5.1.1
trio==0.24.0
trio-websocket==0.11.1
Twisted==22.10.0
typing==3.7.4.3
typing_extensions==4.9.0
urllib3==2.0.7
w3lib==2.1.2
wsproto==1.2.0
zope.interface==6.1
